fix: replace default page with blank page & moved original home page to home.tsx
- this is to make the introduction of the business model when the user first visits the website
